+++ src/ramstk/controllers/__init__.py
@@ -7,7 +7,26 @@
 # Copyright 2007 - 2017 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
 """The RAMSTK controllers package."""
 
-# RAMSTK Local Imports
-from .managers import (
+# RAMSTK Package Imports
+from ramstk.controllers.managers import (
     RAMSTKAnalysisManager, RAMSTKDataManager, RAMSTKMatrixManager
 )
+
+# RAMSTK Local Imports
+from .fmea.analysismanager import AnalysisManager as amFMEA
+from .fmea.datamanager import DataManager as dmFMEA
+from .function.analysismanager import AnalysisManager as amFunction
+from .function.datamanager import DataManager as dmFunction
+from .hardware.analysismanager import AnalysisManager as amHardware
+from .hardware.datamanager import DataManager as dmHardware
+from .hardware.matrixmanager import MatrixManager as mmHardware
+from .options.datamanager import DataManager as dmOptions
+from .pof.datamanager import DataManager as dmPoF
+from .requirement.datamanager import DataManager as dmRequirement
+from .requirement.matrixmanager import MatrixManager as mmRequirement
+from .revision.datamanager import DataManager as dmRevision
+from .stakeholder.analysismanager import AnalysisManager as amStakeholder
+from .stakeholder.datamanager import DataManager as dmStakeholder
+from .validation.analysismanager import AnalysisManager as amValidation
+from .validation.datamanager import DataManager as dmValidation
+from .validation.matrixmanager import MatrixManager as mmValidation
