+++ src/ramstk/controllers/fmea/__init__.py
@@ -6,7 +6,3 @@
 # All rights reserved.
 # Copyright 2007 - 2019 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
 """The RAMSTK FMEA controllers package."""
-
-# RAMSTK Local Imports
-from .analysismanager import AnalysisManager as amFMEA
-from .datamanager import DataManager as dmFMEA
