+++ src/ramstk/controllers/function/__init__.py
@@ -6,7 +6,3 @@
 # All rights reserved.
 # Copyright 2007 - 2019 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
 """The RAMSTK function controllers package."""
-
-# RAMSTK Local Imports
-from .analysismanager import AnalysisManager as amFunction
-from .datamanager import DataManager as dmFunction
