+++ src/ramstk/controllers/requirement/__init__.py
@@ -7,7 +7,3 @@
 # All rights reserved.
 # Copyright 2007 - 2019 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
 """The RAMSTK requirement controllers package."""
-
-# RAMSTK Local Imports
-from .datamanager import DataManager as dmRequirement
-from .matrixmanager import MatrixManager as mmRequirement
