+++ src/ramstk/controllers/stakeholder/__init__.py
@@ -7,7 +7,3 @@
 # All rights reserved.
 # Copyright 2007 - 2019 Doyle Rowland doyle.rowland <AT> reliaqual <DOT> com
 """The RAMSTK stakeholder controllers package."""
-
-# RAMSTK Local Imports
-from .analysismanager import AnalysisManager as amStakeholder
-from .datamanager import DataManager as dmStakeholder
